Improved repetition capabilities
Also updated winput
