+++ winput/__init__.py
@@ -1,0 +1,1 @@
+from .winput import *
